requirements.txt
python-telegram-bot==21.5
flask==3.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.34
python-dotenv==1.0.0
mutagen==1.47.0
werkzeug==3.0.1